Extract commit handler
